Make resque's interval always int
   Redis' BLPOP accepts timeout as integer.

     "The timeout argument is interpreted as an integer value specifying the maximum number of seconds to block"
       -- http://redis.io/commands/blpop

+++ lib/resque/config.rb
@@ -24,7 +24,7 @@ module Resque
     end
 
     def interval
-      @options[:interval].to_f
+      @options[:interval].to_i
     end
 
     def queues
